+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## TBD
+
+### Bug fixes
+
+* Fixed an issue where feature-flags were not always sent if an OnSendCallback was configured
+  [#1589](https://github.com/bugsnag/bugsnag-android/pull/1589)
+
 ## 5.19.1 (2022-01-21)
 
 ### Bug fixes